requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
boto3>=1.34.0
stripe>=7.0.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
cachetools>=5.3.0
